👌 extract test/emulate directory
