+++ packages/core/test/emulate/syntheticsWorkerValues.ts
@@ -1,11 +1,11 @@
-import { deleteCookie, setCookie } from '../src/browser/cookie'
-import type { BrowserWindow } from '../src/domain/synthetics/syntheticsWorkerValues'
+import { deleteCookie, setCookie } from '../../src/browser/cookie'
+import type { BrowserWindow } from '../../src/domain/synthetics/syntheticsWorkerValues'
 import {
   SYNTHETICS_INJECTS_RUM_COOKIE_NAME,
   SYNTHETICS_RESULT_ID_COOKIE_NAME,
   SYNTHETICS_TEST_ID_COOKIE_NAME,
-} from '../src/domain/synthetics/syntheticsWorkerValues'
-import { ONE_MINUTE } from '../src/tools/utils'
+} from '../../src/domain/synthetics/syntheticsWorkerValues'
+import { ONE_MINUTE } from '../../src/tools/utils'
 
 // Duration to create a cookie lasting at least until the end of the test
 const COOKIE_DURATION = ONE_MINUTE
