+++ packages/core/test/index.ts
@@ -1,11 +1,11 @@
 export * from './browserChecks'
 export * from './buildEnv'
 export * from './collectAsyncCalls'
-export * from './createNewEvent'
-export * from './location'
-export * from './mockClock'
+export * from './emulate/createNewEvent'
+export * from './emulate/location'
+export * from './emulate/mockClock'
 export * from './requests'
-export * from './stubReportApis'
-export * from './stubZoneJs'
-export * from './syntheticsWorkerValues'
-export * from './simpleTestDoubles'
+export * from './emulate/stubReportApis'
+export * from './emulate/stubZoneJs'
+export * from './emulate/syntheticsWorkerValues'
+export * from './emulate/simpleTestDoubles'
